Fix some type/format mismatch warnings for pid_t.

+++ base/nagios.c
@@ -161,7 +161,7 @@ static int nagios_core_worker(const char *path)
 		return 1;
 	}
 
-	ret = nsock_printf_nul(sd, "@wproc register name=Core Worker %d;pid=%d", getpid(), getpid());
+	ret = nsock_printf_nul(sd, "@wproc register name=Core Worker %ld;pid=%ld", (long)getpid(), (long)getpid());
 	if (ret < 0) {
 		printf("Failed to register as worker.\n");
 		return 1;
